requirements.txt
streamlit
openai
pandas
python-docx
PyPDF2
openpyxl
python-dotenv
